+++ src/na/components/Header.tsx
@@ -171,6 +171,7 @@ const Header: React.FC = () => {
                       <img
                         src={dd.image}
                         alt=""
+                        loading="lazy"
                         style={{
                           width: '100%',
                           height: '100%',